requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.20
pyodbc==4.0.39
pandas==2.0.3
openpyxl==3.1.2
python-dotenv==1.0.0 